lapr4/1181498-lei_isep_2019_20_sem4_2dd_1181498_1181500_1181255_1181529-3197f42c0ac1/eapli.base/base.core/src/test/java/eapli/base/product/domain/ProductCodeTest.java
package eapli.base.product.domain;

import eapli.base.rawmaterial.domain.RawMaterial;
import eapli.base.utils.Unit;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.jupiter.api.Assertions.*;

public class ProductCodeTest {

    ProductCode instance;

    @Before
    public void before() {
        instance = new ProductCode("1200003", "320030000005");
    }

    /**
     * Test of GetManufacturingCode method, of class ProductCode.
     */
    @Test
    public void testGetManufacturingCode() {
        String exp = "1200003";
        assertEquals(exp,instance.getManufacturingCode());
    }

    /**
     * Test of GetManufacturingCode method, of class ProductCode.
     */
    @Test
    public void testGetCommercialCode() {
        String exp = "320030000005";
        assertEquals(exp,instance.getComercialCode());
    }

}